add get all hotel
